Add file when there isn't any file #348

+++ controllers/object/object_controller.php
@@ -1146,29 +1146,28 @@ class ObjectController extends Controller {
 
 		    	$attachment_id_old = get_post_meta( $data['item_id'],'socialdb_object_content', true);
 
-			    if($attachment_id_old)
+			    $attachment_id = media_handle_upload('new_file', $data['item_id']);
+			    if(!is_wp_error($attachment_id))
 			    {
-				    $attachment_id = media_handle_upload('new_file', $data['item_id']);
-				    if(!is_wp_error($attachment_id))
+			        delete_post_meta($data['item_id'],'socialdb_object_content');
+
+			        if($attachment_id_old) wp_delete_post($attachment_id_old);
+
+				    update_post_meta($data['item_id'], 'socialdb_object_content', $attachment_id);
+				    update_post_meta($data['item_id'], 'socialdb_object_from', 'internal');
+			        if($_FILES['new_file']['type'] == 'application/pdf')
 				    {
-				    	delete_post_meta($data['item_id'],'socialdb_object_content');
-				    	wp_delete_post($attachment_id_old);
-
-					    update_post_meta($data['item_id'], 'socialdb_object_content', $attachment_id);
-				    	if($_FILES['new_file']['type'] == 'application/pdf')
-					    {
-						    $canvas_images[] = ['post_id' => $data['item_id'], 'img' => $data['img']];
-						    save_canvas_pdf_thumbnails($canvas_images, true);
-					    }else
-					    {
-						    set_post_thumbnail($data['item_id'], $attachment_id);
-					    }
-
-					    delete_post_meta( $data['item_id'], '_file_id' );
-					    return true;
-				    }else {
-						return false;
+					    $canvas_images[] = ['post_id' => $data['item_id'], 'img' => $data['img']];
+					    save_canvas_pdf_thumbnails($canvas_images, true);
+				    }else
+				    {
+					    set_post_thumbnail($data['item_id'], $attachment_id);
 				    }
+
+				    delete_post_meta( $data['item_id'], '_file_id' );
+				    return true;
+			    }else {
+					return false;
 			    }
 		    	break;
         }
